+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a C
 
 ## 0.3.0
 
+## Changed
+
+- When storing items with the PSR-16 adapter, the group name is now automatically prefixed to avoid unintended data loss when flushing the cache.
+
 ### Fixed
 
 - Incorrect database table prefix in multisite installations.